pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ainimotion"
version = "0.1.0"
description = "Anime Video Enhancement Pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "torch",
    "torchvision",
    "pillow",
    "scikit-image",
    "numpy",
    "tqdm",
    "pyyaml",
    "tensorboard",
    "wandb"
]

[project.urls]
Homepage = "https://github.com/TWhit229/AInimotion"

# [project.scripts]
# CLI entry point removed — V5 uses scripts/train_v5.py directly

[tool.setuptools.packages.find]
where = ["."]
include = ["ainimotion*"]
